+++ src/main/java/com/vorona/CRUD/Service/UserService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserService {
     public List<User> getUser(/*int count*/);
     public void save (User user);
-    public User show (int id);
-    public void update(int id, User updareUser);
-    public void delete(int id);
+    public User show (long id);
+    public void update(long id, User updareUser);
+    public void delete(long id);
 }
